fix: interval pred
Intervals now do not subtract one day from the end of it. It can be
assumed that adding one month adds one month to the date e.g. if we
start on the first of the month and add one month you will then get the
first of the next month.

If however you would like to have an interval that is a calendar month
you have two choices, either use the CalendarUnit API and iterate by
calendar and then turn it into an interval or apply a RelativeDuration
trick RelativeDuration::month(1).with_days(-1). This will apply a month
movement forward to get you the first of the next month and then apply a
date movement backwards by one day to get you the last day of the month.